+++ monitor/lister_test.go
@@ -1,9 +1,10 @@
-package api
+package monitor
 
 import (
 	"context"
 	"testing"
 
+	"github.com/buildkite/agent-stack-k8s/api"
 	"github.com/stretchr/testify/require"
 	batchv1 "k8s.io/api/batch/v1"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -12,7 +13,7 @@ import (
 	"k8s.io/client-go/kubernetes/fake"
 )
 
-func TestBuildkiteJobManager(t *testing.T) {
+func TestJobLister(t *testing.T) {
 	tag := "some-tag=yep"
 	jobs := []runtime.Object{
 		&batchv1.Job{
@@ -24,8 +25,8 @@ func TestBuildkiteJobManager(t *testing.T) {
 			ObjectMeta: v1.ObjectMeta{
 				Name: "different-tag",
 				Labels: map[string]string{
-					TagLabel:  "something-else",
-					UUIDLabel: "1",
+					api.TagLabel:  "something-else",
+					api.UUIDLabel: "1",
 				},
 			},
 		},
@@ -33,8 +34,8 @@ func TestBuildkiteJobManager(t *testing.T) {
 			ObjectMeta: v1.ObjectMeta{
 				Name: "matching-tag",
 				Labels: map[string]string{
-					TagLabel:  TagToLabel(tag),
-					UUIDLabel: "2",
+					api.TagLabel:  api.TagToLabel(tag),
+					api.UUIDLabel: "2",
 				},
 			},
 		},
